+++ src/main/java/auth/domain/LoginUser.java
@@ -1,4 +1,4 @@
-package nextstep.etc.resolver;
+package auth.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
