+++ src/main/java/nextstep/config/web/interceptor/AdminInterceptor.java
@@ -1,4 +1,4 @@
-package nextstep.etc.interceptor;
+package nextstep.config.web.interceptor;
 
 import auth.domain.UserDetails;
 import auth.provider.JwtTokenProvider;
@@ -7,12 +7,9 @@ import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.Objects;
 
 @RequiredArgsConstructor
-public class JwtInterceptor implements HandlerInterceptor {
-
-    private final String accessTokenName;
+public class AdminInterceptor implements HandlerInterceptor {
 
     private final String loginUserName;
 
@@ -24,12 +21,8 @@ public class JwtInterceptor implements HandlerInterceptor {
             HttpServletResponse response,
             Object handler
     ) {
-        String bearerToken = request.getHeader(accessTokenName);
-        if (Objects.isNull(bearerToken)) {
-            return true;
-        }
-        UserDetails userDetails = provider.getUserDetails(bearerToken);
-        request.setAttribute(loginUserName, userDetails);
+        UserDetails userDetails = (UserDetails) request.getAttribute(loginUserName);
+        provider.validateAdmin(userDetails);
 
         return true;
     }
