+++ src/main/java/nextstep/exception/ReservationException.java
@@ -1,4 +1,6 @@
-package nextstep.etc.exception;
+package nextstep.exception;
+
+import nextstep.exception.message.ErrorMessage;
 
 public class ReservationException extends BaseException {
