+++ src/main/java/nextstep/exception/ScheduleException.java
@@ -1,4 +1,6 @@
-package nextstep.etc.exception;
+package nextstep.exception;
+
+import nextstep.exception.message.ErrorMessage;
 
 public class ScheduleException extends BaseException {
