+++ src/main/java/nextstep/exception/ThemeException.java
@@ -1,4 +1,6 @@
-package nextstep.etc.exception;
+package nextstep.exception;
+
+import nextstep.exception.message.ErrorMessage;
 
 public class ThemeException extends BaseException {
