+++ src/main/java/nextstep/theme/service/ThemeServiceImpl.java
@@ -1,8 +1,8 @@
 package nextstep.theme.service;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.etc.exception.ErrorMessage;
-import nextstep.etc.exception.ThemeException;
+import nextstep.exception.ThemeException;
+import nextstep.exception.message.ErrorMessage;
 import nextstep.theme.domain.Theme;
 import nextstep.theme.dto.ThemeRequest;
 import nextstep.theme.dto.ThemeResponse;
